+++ jenkins-client/src/main/java/com/offbytwo/jenkins/model/BaseModel.java
@@ -14,6 +14,12 @@ import com.offbytwo.jenkins.client.JenkinsHttpClient;
  */
 public class BaseModel {
 
+    private String _class;
+    
+    public String get_class() {
+        return _class;
+    }
+
     //TODO: We should make this private
     protected JenkinsHttpClient client;