+++ jenkins-client/src/main/java/com/offbytwo/jenkins/model/BuildCause.java
@@ -1,3 +1,9 @@
+/*
+ * Copyright (c) 2013 Cosmin Stejerean, Karl Heinz Marbaise, and contributors.
+ *
+ * Distributed under the MIT license: http://opensource.org/licenses/MIT
+ */
+
 package com.offbytwo.jenkins.model;
 
 public class BuildCause {