+++ jenkins-client/src/main/java/com/offbytwo/jenkins/model/QueueItemActions.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2013 Cosmin Stejerean, Karl Heinz Marbaise, and contributors.
+ *
+ * Distributed under the MIT license: http://opensource.org/licenses/MIT
+ */
+
+package com.offbytwo.jenkins.model;
+
+import java.util.List;
+
+public class QueueItemActions extends BaseModel {
+    private List<CauseAction> causes;
+
+    public List<CauseAction> getCauses() {
+        return causes;
+    }
+
+    public void setCauses(List<CauseAction> causes) {
+        this.causes = causes;
+    }
+
+}