+++ jenkins-client/src/main/java/com/offbytwo/jenkins/model/QueueTask.java
@@ -1,6 +1,12 @@
+/*
+ * Copyright (c) 2013 Cosmin Stejerean, Karl Heinz Marbaise, and contributors.
+ *
+ * Distributed under the MIT license: http://opensource.org/licenses/MIT
+ */
+
 package com.offbytwo.jenkins.model;
 
-public class QueueTask {
+public class QueueTask extends BaseModel {
 
     private String name;